🐛 fix: 修复黑体入口、静音记忆反馈及上电低电按键行为
- 测量后5秒等待期间支持长按8秒进入黑体：将 Ready_Timeout 纳入长按检测
- 记忆模式支持长按8秒进入黑体：复用黑体入口，切换后退出记忆处理，松键清零长按计数
- 生产模式进入黑体时清空记忆：在切换模式前调用 Clr_All_Memory
- 静音进入记忆及切换序号时振动一次：在序号显示入口按声音状态启动振动计时
- 非低电首次上电时忽略持续按住的按键：等待测量、设置和记忆键全部释放后再执行主循环
- 低电长按SET后按一次测量键即可关机：进入低电错误态时预置开关键释放标志
- 保持按键功能和高精度显示不变并减少ROM占用：合并相同按键分支，将数字临时变量缩为uint8

+++ Src/Drv_LCD_ET29A.c
@@ -368,7 +368,7 @@ void Disp_High_Precision(uint8 Unit, int16 Temp)
 {
     if(High_Precision_Flag)
     {
-        uint16 R_LCD1,R_LCD2,R_LCD3,R_LCD4,R_LCD5;
+		uint8 R_LCD1,R_LCD2,R_LCD3,R_LCD4,R_LCD5;
         
 
         //强制转换成无符号数
